+++ gems/ast-crispr/spec/ast/crispr/version_spec.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require "anonymous_loader"
+require 'anonymous_loader'
 RSpec.describe Ast::Crispr::Version do
-  it_behaves_like "a Version module", described_class
+  it_behaves_like 'a Version module', described_class
 
-it "executes the version file for coverage without redefining constants" do
-  path = File.expand_path("../../../lib/ast/crispr/version.rb", __dir__)
-  anonymous_namespace = AnonymousLoader.load(files: path)
+  it 'executes the version file for coverage without redefining constants' do
+    path = File.expand_path('../../../lib/ast/crispr/version.rb', __dir__)
+    anonymous_namespace = AnonymousLoader.load(files: path)
 
-  expect(anonymous_namespace::Ast::Crispr::Version::VERSION).to eq(described_class::VERSION)
-end
+    expect(anonymous_namespace::Ast::Crispr::Version::VERSION).to eq(described_class::VERSION)
+  end
 end